+++ modules/core/unit/settings/storage_duration.cpp
@@ -148,7 +148,7 @@ NT2_TEST_CASE( automatic_duration_apply )
                                 >::type
                               >
                       )
-                    , (array_buffer<int, boost::mpl::integral_c<long,4> >)
+                    , (array_buffer<int, boost::mpl::integral_c<size_t,4> >)
                     );
 
   NT2_TEST_EXPR_TYPE( automatic_()
@@ -163,7 +163,7 @@ NT2_TEST_CASE( automatic_duration_apply )
                                 >::type
                               >
                       )
-                    , (array_buffer<int, boost::mpl::integral_c<long,36> >)
+                    , (array_buffer<int, boost::mpl::integral_c<size_t,36> >)
                     );
 }
